move computation of interior point triangulation to MeshCell
